MB-13234: Create RemoteHelper only if SSH credentials are given
If there are no SSH credentials given in the cluster configuration,
no RemoteHelper will be created. That means that some setup steps
that need remote access (like dropping caches) won't be performed.

The reason for this change is to make it possible to configure a
perfrunner cluster without SSH access. This is the case when you
simulate a cluster locally with ./cluster_run.

Change-Id: I561b78badd01cf1423273a05ec14770686fac240
Reviewed-on: http://review.couchbase.org/46000
Reviewed-by: thomas anderson <thomas.couchbase@gmail.com>
Tested-by: thomas anderson <thomas.couchbase@gmail.com>