Add the ability to create n1ql indexes during tests
This change allows us to create primary and secondary indexes
based on parameters set in the test config files. In order to
add indexes through n1ql we have added the [n1ql] section to
the config file which will be used for n1ql setup. Currently
the only thing that can be specified in this section is the
indexes that you want created. Only one parameter can be
specified in this section.

indexes - This parameter will take a list of index creation
statements to be run during cluster setup. In places where
a bucket name should be specified in the query just add
{bucket} and perfrunner will fill in the appropriate name.

Lastly, this change adds the test setup for the N1QL PM1 test
specification. This is currently only a partial specification.
The cluster is set up properly and loaded with items, but the
access phase is not complete yet. Completion of the access phase
will happen in a future change.

Change-Id: I2d6bef3e7f6b8c9b2073e524ab8fa22493fc3077
Reviewed-on: http://review.couchbase.org/49929
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Michael Wiederhold <mike@couchbase.com>