MB-13234: Add cluster spec for local vagrant VMs
If you want to run perfrunner locally, you can spin up vagrant machines [1].
As every version/OS is on different IPs, they are not specified in the configuration
file. Instead they are supplied as command line parameter like that:

    ./env/bin/python -m perfrunner.utils.cluster -c clusters/vagrant.spec -t tests/<your-test> \
    "clusters.vagrant.\"192.168.109.101:8091 192.168.109.102:8091 192.168.109.103:8091 192.168.109.104:8091\""

The example above would be an a cluster with Ubuntu 12.04 running the Couchbase 4.0
developer preview.

[1]: https://github.com/couchbaselabs/vagrants

Change-Id: If076bba81021509967b31d0d35051c24a52b1b7c
Reviewed-on: http://review.couchbase.org/51624
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Volker Mische <volker.mische@gmail.com>
Tested-by: Volker Mische <volker.mische@gmail.com>