Add support for multiple indexer nodes to SecondaryIndexTest base class for both initial and incremental workload.
Pass in modified index names for partitioned index when waiting for
incremental index build. Also pass in all index nodes to fetch stats on
each index node.

When waiting for incremental index build to complete, allow it to pass
if the count is within a percentage of the expected AND if the number
has not changed for 20 loops (seconds).

Add support for multiple indexer nodes to the helpers.rest's
wait_for_secindex_init_build . Also pass in the list of active indexes.
Active indexes are just indexes with suffixes to represent the
partitions of indexes.

Tested the following scenario:

shortened version of secondary_20M_multiple.test with 2 indexes/fields
and no partitioning of indexes over indexer nodes. This is the backward
compatible case.

Multiple indexers test: 2 indexes/fields over 2 indexer nodes.

Change-Id: I4dbc83c5d10717d602ba9e04dd229dedfba0b7a6
Reviewed-on: http://review.couchbase.org/52002
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: David C Kao <dkao_cb@yahoo.com>