Simplify test settings
Instead of defining some attributes multiple times, defined them
only once in their designated section. For example `size` and `doc_gen`
is now defined *only* in the `[load]` section, but re-used in the `[hotload]`
and `[access]` section.

An error will be thrown if `size` is used in `[access]` or `[hotload]`.

Change-Id: I563c02491e98f3302f1ce3a00081c72eb890ae75
Reviewed-on: http://review.couchbase.org/52117
Reviewed-by: Volker Mische <volker.mische@gmail.com>
Tested-by: buildbot <build@couchbase.com>