Add section information to settings
All settings now contain a `_section` property which has the section name
as value. This makes it possible to determine e.g. whether you are in the
access or the load phase.

From now on properties with a leading underscore are ones that are not
specified in the configuration file, but are generated in some other ways.

Change-Id: I437f39abd51be6007c80460e7a82aece8c16c7cf
Reviewed-on: http://review.couchbase.org/52831
Reviewed-by: Dave Rigby <daver@couchbase.com>
Reviewed-by: thomas anderson <thomas@couchbase.com>
Tested-by: buildbot <build@couchbase.com>