CBQE-3242: Add a run start time as a run identifier
Background: perf daily sanity has 2 different data sources, one is the perfrunner
based code and the other is query benchmark. Both write the results to a couch base.
In order to make the results of these runs commonly query able they need common
fields, one of which is the build and the other will be a start time passed by the
calling Jenkins job. Note that build is in itself is not sufficient because we may
do multiple runs on the same build.

The _beta in the file name mean this is the development version, once it runs well
in production it will be copied to the main script.

Change-Id: Iface32a2255e16b7fe5077aac5bdd00854c1958f
Reviewed-on: http://review.couchbase.org/58906
Reviewed-by: Eric Cooper <ericcouchbase@gmail.com>
Tested-by: Eric Cooper <ericcouchbase@gmail.com>