Use the same option for both build versions and custom URLs
This is what many Jenkins jobs are currently doing:

  if [ -z "${url}" ]
  then
    env/bin/install -c ${cluster} -v ${version}
  else
    env/bin/install -c ${cluster} --url ${url}
  fi

It seems annoying that we need to modify our jobs in order to
support different inputs. Computers are smart enough to find the
difference between build numbers and links.

That becomes even more critical because the build team has removed
4.5.0, 4.5.1, and 4.6.0 builds. We have to use direct links when
we need to test those releases.

The change is backward-compatible.

Change-Id: Ib6665170b27391afacd898a3ee475f260ae09841
Reviewed-on: http://review.couchbase.org/76222
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Pavel Paulau <pavel.paulau@gmail.com>