Allow to load documents in unsorted order
Currently, perfrunner loads all document in a sorted order.

Each key consists of a prefix and a numerical counter, e.g.:

  "38d7cd-000072438963"

This is how per-worker counters are incremented when 10 workers
are loading 100 documents:

  W1  :  1, 11, 21, 31, 41, 51, 61, 71, 81, 91
  W2  :  2, 12, 22, 32, 42, 52, 62, 72, 82, 92
  ...
  W10 : 10, 20, 30, 40, 50, 60, 70, 80, 90, 100

The sequential counter is later converted to a 12 characters long
string:

  key = "%012d" % n

Essentially, all documents are sorted by that counter.

This patch introduces two important changes. First of all, each
worker owns a different set of keys now:

  W1  :  1,  2,  3,  4,  5,  6,  7,  8,  9, 10
  W2  : 11, 12, 13, 14, 15, 16, 17, 18, 19, 20
	...
  W10 : 91, 92, 93, 94, 95, 96, 97, 98, 99, 100

Also, the order of those numbers is randomized using a rotating
sequence with a prime number (971 is chosen):

	n = (n + prime) % total_keys

Here is an example of generated numbers for 10000 documents:

  972, 1943, 2914, 3885, 4856, 5827, 6798, 7769, 8740, 9711, 682, 1653, ...

This is a simple and yet fast way to generate pseudo-random
sequences of keys.

Note, each worker is an OS process. Unfair scheduling, resource
contention, and variable latency of database operations are an
additional jitter.

Change-Id: I69af3e58c488b3eade1554fae82688654975919b
Reviewed-on: http://review.couchbase.org/77069
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Pavel Paulau <pavel.paulau@gmail.com>