Get rid of host_port notation
When perfrunner was created we wanted to make it compatible with
cluster_run. It was decided to specify ns_server ports in cluster
specifications (8091 or 9000).

Eventually this compatibility was completely broken. Also we
started using HTTP services with other ports (GSI, N1QL, FTS).

There is no point to keep using 8091 as the primary port in
in perfrunner automation. The idea is to use just the hostname
throughout the source code and add custom port when needed.

This change should also help to start using Ansible inventories in
perfrunner.

Change-Id: I3a2b15ad28030c6ee4120af4c6f8d89cb1bc120b
Reviewed-on: http://review.couchbase.org/78589
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Pavel Paulau <pavel.paulau@gmail.com>