CBPS-339 merge conflict resolution
This project is about integrating perfrunner with cloud.
In this patch AWS instances are started using apache libcloud,
When we launch the instances we store the information of
instance_id, public ip address. The instance_id will be used to
track the instances for deletion. The AMIs are created
optimized for server and client. The client is ubuntu
so to login you need to use ubuntu as user, ec2-user
for the server. Server is craeted with EBS volume
attached, which will be deleted  after nodes are destroyed.
to run: perfrunner.utils.cloudrunner -p start -k useme

Change-Id: I808805989f5c6e1ef0e90368c272031d5d7e4c7e
Reviewed-on: http://review.couchbase.org/79910
Reviewed-by: sandip nandi <sandip.nandi@couchbase.com>
Tested-by: sandip nandi <sandip.nandi@couchbase.com>