CBPS-339 Adding files for cloud
This project is about integrating perfrunner with cloud.
In this patch AWS instances are started using apache libcloud,
When we launch the instances we store the information of
instance_id, public ip address. The instance_id will be used to
track the instances for deletion. The AMIs are created
optimized for server and client. The client is ubuntu
so to login you need to use ubuntu as user, ec2-user
for the server. Server is craeted with EBS volume
attached, which will be deleted  after nodes are destroyed.
to run: perfrunner.utils.cloudrunner -p start -k useme

Change-Id: I88d15178ee36a734ed2930827df9d1db425bd174
Reviewed-on: http://review.couchbase.org/78810
Reviewed-by: sandip nandi <sandip.nandi@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>