Use installer's artifacts in test cases with tools
Currently, our performance tests are not aware of the packages used
during the installation. The tests use REST API to detect the actual
version of Couchbase Server and then try to find the corresponding
package using one of common locations.

Although that works just fine for the release and development builds,
there is no way we can match arbitrary toy builds.

It's probably OK if "install" script saves the package locally and
the test extracts it later during the execution.

Change-Id: Ieda23ff22856fb0b4b1423101d5270165c3212d8
Reviewed-on: http://review.couchbase.org/80334
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Pavel Paulau <pavel.paulau@gmail.com>