Collect logs before restarting the servers
Usually we run cbcollect_info after the test completion. However,
servers have to be restarted at the end of the test when custom
memory settings are used. Many kv-engine and ns_server stats are
not persisted and hence being lost.

We should collect the logs before resetting the memory settings.

Change-Id: I554f1f2f06499e12560d226b3c5c908c3dde09d0
Reviewed-on: http://review.couchbase.org/85204
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Pavel Paulau <pavel.paulau@gmail.com>