CBPS-622 Make optional backup flags configurable in test spec
This commit makes optional flags for cbbackupmgr configurable
under the [backup] group in the test spec. This enables greater
flexibility when creating backup tests.

The default behaviour is not to inlclude these flags if these
test paramaters aren't included. As a result, exisitng tests have
been updated to specify the thread count as 16 (to be consistent
with all previous test runs).

As cbbackupwrapper will be reviewed at a later date, its default
for threads has been hard coded as 16 (for now), should the
parameter not be provided.

Change-Id: I51c10b0b813abb54c6df48675c4e77bc0cdedd43
Reviewed-on: http://review.couchbase.org/107470
Reviewed-by: Patrick Varley <patrick@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>