CBPS-630 Merge test improvements
This commit removes the flush between backups to give
a more realistic backup/merge scenario.
This also adds configurable threads and storage flags
for merge. The threads flag will be ignored for builds
pre 6.5.0-3198, where this was introduced. Setting the threads
to the number that is used by backup should maintain conistency
with previous tests, due to the way thread count was determined before.

Change-Id: I30b5d28da6b3c37bd5959c65d6a06bff0303fe84
Reviewed-on: http://review.couchbase.org/109205
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Patrick Varley <patrick@couchbase.com>