CBPS-636 Add incremental backup tests
This commit adds incremental backup tests. It does this by running
two load phases, one for initial load, and one borrowing the 'creates'
field to specify a number of upserts with a different key prefix,
creating an additional dataset. The access phase then handles
deletes and upserts on the intial dataset. This is calulated to be
10% under the observation that the number of operations performed
is (ops * workers * time).

Change-Id: Iaa2929ef928f1e5a61f272d8763720a701083632
Reviewed-on: http://review.couchbase.org/109053
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Patrick Varley <patrick@couchbase.com>