MB-36826: DCPThroughputTest: Compact bucket before access phase
Prior to this change large variations are observed in the DCP stream
runtime as it depends significiantly on exactly how fragmented the
bucket is when the access phase starts. Given the test disables
compaction during the load phase, code changes which improve the
efficiency of the flusher (hence writing in quicker, smaller batches)
can result in _worse_ metrics as DCP backfilling requires reading from
a larger, more fragmented file.

To ensure consistent results for the DCP throuhgput phase of the test;
ensure the Bucket is compacted before measuring how long it takes to
stream its contents via DCP.

Change-Id: Ie1d8177ff6454460fbcbf8c8aa6a12f98c168f2f
Reviewed-on: http://review.couchbase.org/117650
Reviewed-by: Korrigan Clark <korrigan.clark@couchbase.com>
Tested-by: Korrigan Clark <korrigan.clark@couchbase.com>