CBPS 910: no debug symbols for toy builds
When linux perf profiling is enabled for a test, we download a debuginfo rpm for the installed Couchbase Server version. Previously, we couldn't handle the case where a toybuild server is used as we install the server and debuginfo in separate scripts (utils/install.py and utils/cluster.py resp.).

This commit moves the installation of the debug symbols into install.py, using the URL of the Couchbase Server package it will install to derive the URL for the matching debug symbols.

Change-Id: I1748f6646bbfac4ccaa44848bbe5b079a6d4ecd8
Reviewed-on: https://review.couchbase.org/c/perfrunner/+/163773
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Salim Salim <salim.salim@couchbase.com>