Add N1QL YCSB latency tests
There was no working test or results on showfast for N1QL YCSB workloade
latency, so I took the equivalent test from the SDK tab, added versions
for 1 node and plasma. I also run the test on showfast for multiple
versions of couchbase server. Finally, I think it is a good idea to run
latency tests weekly, so I added them to the weekly pipeline.

Change-Id: I7ad1a6e80a38e680598d2e50d660624cad4ad9df
Reviewed-on: http://review.couchbase.org/c/perfrunner/+/164449
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Korrigan Clark <korrigan.clark@couchbase.com>