Update SDK version requirements
Starting from Couchbase Server 7.1.0-1745, only C and Python SDK versions >= 3.2.0 are fully supported. See https://issues.couchbase.com/browse/MB-45563

This change updates the minimum SDK version requirements for many tests. Specifically, if a test doesn't specify any C or Python SDK AND the test is on Couchbase Server >= 7.1.0-1745, we install 3.2.3 by default. For older Couchbase Server builds we install 3.0.8 by default (this is to minimise the impact of CCBC-1515 for now). We also update requirements.txt to set the default version of the python SDK to 3.2.3 - this is required for remote clients, who do not run clients.py and so rely on the version of the python SDK defined in requirements.txt.

Change-Id: I0657347ef59e458ba7a00475acbfd7bf2238de9a
Reviewed-on: https://review.couchbase.org/c/perfrunner/+/166223
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Bo-Chun Wang <bo-chun.wang@couchbase.com>