Reduce load workers in daily YCSB workload A test
Daily workload A tests were often failing as they were not loading the correct number of documents. Lowering the number of workers or increasing the number of available CPU cores on the server solved this issue. To keep the test consistent the load workers are the best thing to change.

Change-Id: I802aada3bcd866c3fff2fbda2ece3775414ec39d
Reviewed-on: https://review.couchbase.org/c/perfrunner/+/171481
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Bo-Chun Wang <bo-chun.wang@couchbase.com>