Disable access log creation in warmup tests
Warmup happens slightly differently depending on whether there is an
access log present. The access log is generated at 2am GMT every day -
currently, it could happen (and has happened) that an access log is
created during the test, before the warmup phase. This then causes the
warmup phase to behave slightly differently to how we want it.

This patch disables access log creation for warmup tests to avoid this.

Change-Id: I1c5dca0817878863c0c11aa1b4166b61838516a9
Reviewed-on: https://review.couchbase.org/c/perfrunner/+/180605
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dragos Petru Taraban <dragos.petru@couchbase.com>