Use AMI names instead of IDs for AWS
Keeping track of AMI IDs in our deployment scripts is a pain. Its
better for us to use AMI names in our scripts, and use Terraform's Data
Sources to lookup the correct AMIs to use based on the name.

This patch is a simple implementation of this which just looks up the
exact AMI name that is in terraform.py.

In the future, this can be made more sophisticated, e.g. to be able to
lookup "client-ami-*" with a wildcard, and selecting the most recent one
(this can be done using Terraform).

Change-Id: Iee03eb2941cc62819ca0dc31ca6f5045b354ac31
Reviewed-on: https://review.couchbase.org/c/perfrunner/+/181279
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: vikas chaudhary <vikas.chaudhary@couchbase.com>