Fix 'get_nodes' method of Collector to work for cloud tests
This issue taught me something I didn't know about generator functions in Python. Specifically, that when you use 'return' in a generator function, it simply stops the generator.

This meant that in 'get_nodes', if we had self.cloud_enabled = True then we wouldn't return the full list of hostnames as we wanted to, but would rather just return an empty generator.

Instead of using 'return', we need to yield each hostname.

Change-Id: I2d71b71cc0cd3a6ba0f491d97b712a1303b5a897
Reviewed-on: https://review.couchbase.org/c/perfrunner/+/181476
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: vikas chaudhary <vikas.chaudhary@couchbase.com>