Support KV workloads with mixed durability levels
The goal here was to run a KV test with a single bucket where some
writes are durable and some aren't.

The way chosen to achieve this was to update the existing system of
doing a "mixed workload".

Prior to this patch, one could set up a "mixed workload" multi-bucket
test, where different buckets could be targetted with different
workloads. There was no support for having multiple different workloads
happening in the same bucket.

This patch changes that by changing the way that target buckets are
specified in mixed workload settings:
 - Instead of specifying `num_buckets` for each workload in a mixed
 workload, we now specify `bucket_numbers`, which is a comma-separated
 list of bucket indexes to target.
 - E.g. if I want one part of my mixed workload to target the first 3
 buckets, I can do `bucket_numbers = 1,2,3`
 - `bucket_numbers` is used to figure out the list of target buckets for
 each part of a mixed workload. By default, it will be all the available
 buckets. This way, you can do a mixed workload in a single bucket test.

Change-Id: I15e74a14128c4d507b7735367c121b034fe8431b
Reviewed-on: https://review.couchbase.org/c/perfrunner/+/189068
Reviewed-by: Salim Salim <salim.salim@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>