Be able to only deploy capella cluster with terraform.py
Adds two flags that allow one to use the terraform.py script in order to
only deploy the capella cluster (serverless dataplane or provisioned
cluster), without deploying client/utility
nodes or serverless dbs.

This can be useful for debugging sometimes or doing some kind of manual
work on the cluster or dataplane, or if one sets up 2 Jenkins jobs: 1 to deploy a
cluster or dataplane and another to run a test on it (and then keep it around to
run more tests).

Change-Id: Ie517905a67e7f0d5c89e2ff6a12070b181f7ac30
Reviewed-on: https://review.couchbase.org/c/perfrunner/+/190191
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Laura Silaja <laura.silaja@couchbase.com>