Make avg_ops metric sum up ops/sec over buckets and start documenting metrics
Previously, the `avg_ops` metric would give a bucket-level average
ops/sec which is a slightly strange metric which is of dubious value in
multi-bucket tests.

Example of previous behaviour:
- Set up: 2 buckets, one doing exactly 1000 ops/sec and the other
doing exactly 2000 ops/sec for the same amount of time
- avg_ops would give (1000 + 2000) / 2 = 1500

This patch changes the calculation to similar to the `max_ops` metric, so that in this case by default we
would get an average _total_ ops/sec of 3000.

All weekly tests using `avg_ops` be unaffected since they use 1 bucket
only.

Additionally, this patch adds some documentation to throughput metrics
and makes it possible to specify a subset of buckets to `avg_ops` and
`max_ops` for calculations.

Change-Id: Ia3d815be525d591c2a728a9f835415f0541c9b3a
Reviewed-on: https://review.couchbase.org/c/perfrunner/+/194379
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Daniel Nagy <daniel.nagy@couchbase.com>