Refactor get_cbstats function and its callers
Motivation:
I noticed lots of errors like this: "Unexpected exception in KVStoreStats: 'bool' object has no attribute 'split'" in some magma tests and while investigating I saw that the get_cbstats function would either return the output of a `local(..., capture=True)` call, or a bool literal False in case the cbstats command failed. We do not however handle the False case at all.

Refactor:
This change refactors the get_cbstats function and its callers.
Specifically it does this:

1. Rename the function to "run_cbstats" because we also use it to reset
stats, not just to get stats.

2. Changes the function signature so that we pass username/password
instead of a ClusterSpec object, making it more explicit what the
function needs.

3. Adds `bucket` argument to the function signature so we can specify a
bucket for the cbstats command.

4. Refactors where the function is called so that we just get stats for
a specific bucket that we want, instead of getting stats for all buckets
and then filtering the output.

5. Uses `run_local_shell_command` instead of fabric's `local` function
in order to actually run cbstats. Like this we get logging automatically
in the case of errors.

Change-Id: I8f2391f6174c7977c9025d7830186c60fb55c435
Reviewed-on: https://review.couchbase.org/c/perfrunner/+/201226
Reviewed-by: Salim Salim <salim.salim@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>