Fix initial rebalance for cluster setup when there are inactive clusters
Until now, if we have any "inactive clusters" in the cluster spec then
we don't correctly adjust the initial_nodes test config setting and so
the active clusters don't necessarily get set up correctly in the
initial rebalance.

This fixes that by only using the values in initial_nodes that
correspond to active clusters, when we do the initial rebalance.

Change-Id: If178baf388827e7634b9b152d4f1ecc42aa0b1ca
Reviewed-on: https://review.couchbase.org/c/perfrunner/+/205571
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dragos Petru Taraban <dragos.petru@couchbase.com>