AV-73707: add test configs, cluster specs + analytics config refactoring
Changed how analytics configuration REST API settings are handled. Now
we can specify any valid setting in the test config and have it applied
during cluster set up, instead of having to have an explicit property
defined in AnalyticsSettings for each setting.

This means that we are unrestricted in what analytics config REST API
settings we can tweak in tests.

Change-Id: I72a1418d8208df0be1052444603c185fe561de5b
Reviewed-on: https://review.couchbase.org/c/perfrunner/+/206644
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Salim Salim <salim.salim@couchbase.com>