CBPS-1450: only update query settings whose values have changed
Previously we would get all the query settings as a JSON object, update some of the values, and then POST the entire JSON object to update the query settings. So even though we only changed a few of the settings, we are "updating" all of them.

This used to be fine, but from 7.6.3 onwards this triggers MB-63518 and causes tests to get stuck.

Change-Id: I92381b2d1a321b04e311e46531824a5ea2598d4c
Reviewed-on: https://review.couchbase.org/c/perfrunner/+/216092
Reviewed-by: Salim Salim <salim.salim@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>