CBPS-1492: specify allowed IPs using env var
Adds an `allowed_ips` input variable to our terraform scripts, which denotes the IPs/CIDRs that are allowed to SSH into cloud VMs.

`allowed_ips` is configured within terraform.py and can be set using the env var PERF_ALLOWED_IPS as a comma-separated list of IPs/CIDRs. It is generally expected that these are all non-overlapping addresses.

The public IP address of the local machine is always included in `allowed_ips`.

Change-Id: Id659653d39fd383e654460f31ef4153400ad8c20
Reviewed-on: https://review.couchbase.org/c/perfrunner/+/221589
Reviewed-by: Salim Salim <salim.salim@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>