CBPS-1379: minimise broker node configurability
Credit to salim.salim@couchbase.com for the main idea here.

Only broker node instance type can be configured via a "profile" setting in the cluster spec. Broker nodes are no longer able to be deployed with attached data disks either, since the extra space is not required and we don't mount the disks in tests anyway.

The motivation here is to avoid wasting money on broker nodes.

Change-Id: If4faa2a303aa7661946f30b7c90de310a70cf4f4
Reviewed-on: https://review.couchbase.org/c/perfrunner/+/223774
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Salim Salim <salim.salim@couchbase.com>