Fix how passwords are passed to external python scripts
Python's ArgumentParser has problems parsing input of the form "--flag
'-some string starting with a hyphen'". In this case the string starting
with the '-' character is interpreted as a CLI flag, and we get an error
about `--flag` not having a value.

To fix this, we can do the following instead: "--flag='-some string
starting with a hyphen'".

Change-Id: I487c397d003fc955342b76498f391d39945ef5ce
Reviewed-on: https://review.couchbase.org/c/perfrunner/+/231579
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Bo-Chun Wang <bo-chun.wang@couchbase.com>