Avoid running query phase in BigFUN sync-only test cases
Previously, we would run the BigFUN query phases (warmup and access) in
all BigFunInitialSyncAndQuery tests, relying on the default values of
the access phase parameters to avoid actually running queries in
sync-only tests.

This didn't work as the default value for `TestConfig.access_settings.ops` is
`float("inf")` which we try to cast to an int.

+ Add a guard in `BigFunInitialSyncAndQuery.run` to skip the query
  phases if `TestConfig.access_settings.workers` is < 1 (which is the
  default if not specified).

Change-Id: Ic707354ae10249279498118b1c8126406ea879fb
Reviewed-on: https://review.couchbase.org/c/perfrunner/+/234431
Reviewed-by: Salim Salim <salim.salim@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>