Fix reported build in on-prem analytics tests
Change 235150 inadvertently broke the build string for on-prem analytics
tests, resulting in "0.0.0 : " being prepended to the couchbase build.

The logic for checking if the analytics nodes use a different
build/version than other cluster nodes was incorrect when a single
cluster with KV nodes and separate CBAS nodes is used (i.e. most on-prem
analytics tests).

Change-Id: I72e1e252d29230d99dadddc947dc088978f98841
Reviewed-on: https://review.couchbase.org/c/perfrunner/+/237085
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Salim Salim <salim.salim@couchbase.com>